django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-oauth-toolkit>=3.2.0",
    "django-redis>=6.0.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.2",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.0",
    "transformers>=4.57.6",
]
